Added Multiple levels
Added level 5 and 6, the base levels for level 3 and 4 - to be developed.

Added a trigger box to level 2 so the player can progress to level 3.

Added an end screen that plays when the player finishes level 6.

Added a menu that is the default level when the game is not in the editor